ide: Fixed problem with Remove found lines, CodeEditor::GotoLine renamed to GotoBarLine, Navigator now shows number of items in the nest

+++ uppsrc/CtrlLib/LineEdit.cpp
@@ -1216,7 +1216,7 @@ void LineEdit::DeleteLine()
 	}
 	int i = GetLine(cursor);
 	int p = GetPos32(i);
-	Remove((int)p, GetLineLength(i) + 1);
+	Remove(p, GetLineLength(i) + 1);
 	PlaceCaret(p);
 	Action();
 }
